Iris/requirements.txt
django >=1.9, <2
matplotlib
pillow
opencv-python
requests